create controllers, put all db calls in DBConnection
