+++ SourceCode/RootProject/Project/src/main/java/Controllers/BillingController.java
@@ -1,0 +1,14 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Controllers;
+
+/**
+ *
+ * @author Tyler
+ */
+public class BillingController {
+    
+}
